Move singer display class into the singer list because there's no other locaiton user this class.
